refactor(core): remove exports from barrel files
