+++ packages/core/src/index.ts
@@ -9,16 +9,14 @@ export { default as Handle } from './components/Handle/Handle.vue'
 
 export { default as Panel } from './components/Panel/Panel.vue'
 
-export {
-  StraightEdge,
-  StepEdge,
-  BezierEdge,
-  SimpleBezierEdge,
-  SmoothStepEdge,
-  BaseEdge,
-  EdgeText,
-  EdgeLabelRenderer,
-} from './components/Edges'
+export { default as StraightEdge } from './components/Edges/StraightEdge'
+export { default as StepEdge } from './components/Edges/StepEdge'
+export { default as BezierEdge } from './components/Edges/BezierEdge'
+export { default as SimpleBezierEdge } from './components/Edges/SimpleBezierEdge'
+export { default as SmoothStepEdge } from './components/Edges/SmoothStepEdge'
+export { default as BaseEdge } from './components/Edges/BaseEdge.vue'
+export { default as EdgeText } from './components/Edges/EdgeText.vue'
+export { default as EdgeLabelRenderer } from './components/Edges/EdgeLabelRenderer.vue'
 
 export {
   getBezierPath,
@@ -43,7 +41,7 @@ export {
   getRectOfNodes,
   pointToRendererPoint,
   rendererPointToPoint,
-  /** @deprecated will be removed in the next major version, use `rendererPointToPoint` instead */
+  /** @deprecated - will be removed in the next major version, use `rendererPointToPoint` instead */
   rendererPointToPoint as graphPosToZoomedPos,
   getNodesInside,
   getMarkerId,
@@ -53,12 +51,13 @@ export {
 } from './utils/graph'
 
 /**
+ * @deprecated - Use store instance and call `applyChanges` with template-ref or the one received by `onPaneReady` instead
  * Intended for options API
  * In composition API you can access apply utilities from `useVueFlow`
  */
 export { applyChanges, applyEdgeChanges, applyNodeChanges } from './utils/changes'
 
-export { defaultEdgeTypes, defaultNodeTypes } from './store/state'
+export { defaultEdgeTypes, defaultNodeTypes } from './utils/defaultNodesEdges'
 
 export { VueFlow as VueFlowInjection, NodeId as NodeIdInjection } from './context'
